+++ challenge/models.py
@@ -1,7 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import User
+import hashlib # Import hashlib
+from django.core.exceptions import ValidationError  # Import ValidationError
 
-# Create your models here
 class Challenge(models.Model):
     id = models.AutoField(primary_key=True, unique=True)
     name = models.CharField(max_length=100, unique=True)
@@ -18,11 +19,13 @@ class Challenge(models.Model):
     def __str__(self):
         return self.name
     
-    # overwriting default save  method
+    # Overriding default save method
     def save(self, *args, **kwargs):
         if self.start_port > self.end_port:
-            raise Exception("Start port should be less than end port")
-        # Here flag need to be hashed
+            raise ValidationError("Start port should be less than end port") # Raise ValidationError if start_port is greater than end_port
+        if self.flag:
+            if not self.flag.startswith("hashed_"):
+                self.flag = "hashed_" + hashlib.sha256(self.flag.encode('utf-8')).hexdigest()
         super(Challenge, self).save(*args, **kwargs)
 
 class UserChallenge(models.Model):
@@ -40,7 +43,6 @@ class UserChallenge(models.Model):
     is_live = models.BooleanField(default=False)
     no_of_attempt = models.IntegerField(default=0)
     is_solved = models.BooleanField(default=False)
-    port = models.IntegerField()
     
     def __str__(self):
         return f"{self.user.username} - {self.challenge.name}"